First pass impl of backcompat new logger api
