+++ common/src/main/java/com/microsoft/identity/common/internal/logging/RequestContext.java
@@ -22,18 +22,6 @@
 // THE SOFTWARE.
 package com.microsoft.identity.common.internal.logging;
 
-import com.google.gson.Gson;
-
-import java.util.HashMap;
-
-// TODO I'm not wedded to this name, but the concept may work for tracking correlationIds
-public class RequestContext extends HashMap<String, String> implements IRequestContext {
-
-    private static final long serialVersionUID = -1128593786133166935L;
-    private final Gson mGson = new Gson();
-
-    @Override
-    public String toJsonString() {
-        return mGson.toJson(this);
-    }
+public class RequestContext extends com.microsoft.identity.common.logging.RequestContext {
+    // backcompat
 }
