+++ common/src/main/java/com/microsoft/identity/common/logging/ILoggerCallback.java
@@ -20,8 +20,20 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-package com.microsoft.identity.common.internal.logging;
+package com.microsoft.identity.common.logging;
 
-public interface IRequestContext extends com.microsoft.identity.common.logging.IRequestContext {
-    // Class exists for backcompat
+/**
+ * Interface for apps to configure the external logging and implement the callback to designate the
+ * output of the log messages.
+ */
+public interface ILoggerCallback {
+    /**
+     * Interface method for apps to hand off each log message as it's generated.
+     *
+     * @param tag         The TAG for the log message.
+     * @param logLevel    The {@link Logger.LogLevel} for the generated message.
+     * @param message     The detailed message.
+     * @param containsPII True if the log message contains PII, false otherwise.
+     */
+    void log(String tag, Logger.LogLevel logLevel, String message, boolean containsPII);
 }
